+++ hotel/src/app/components/mobileNav.js
@@ -1,19 +1,19 @@
 'use client'
 
 import { useState, useEffect } from "react";
-import styles from ".././main.css"
+import styles from "../main.css";
 
-const ToggleComponent = () => {
+const ToggleComponent = ({ isLoggedIn, onLogout }) => {
     const [isVisible, setIsVisible] = useState(false);
     const [isAnimating, setIsAnimating] = useState(false);
 
     const toggleVisibility = () => {
         if (isVisible) {
-            setIsAnimating(true); // Start the slide-out animation
+            setIsAnimating(true);
             setTimeout(() => {
-                setIsVisible(false); // Remove from DOM after animation
-                setIsAnimating(false); // Reset animation state
-            }, 375); // Match this duration with the CSS animation duration
+                setIsVisible(false);
+                setIsAnimating(false);
+            }, 375);
         } else {
             setIsVisible(true);
         }
@@ -21,37 +21,69 @@ const ToggleComponent = () => {
 
     useEffect(() => {
         document.body.style.overflow = isVisible ? 'hidden' : 'auto';
-
         return () => {
             document.body.style.overflow = 'auto';
         };
     }, [isVisible]);
 
+    // Reset menu visibility if logged out
+    useEffect(() => {
+        if (!isLoggedIn && isVisible) {
+            setIsVisible(false);
+        }
+    }, [isLoggedIn]);
+
+    // Define menu items dynamically based on logged-in state
+    const menuItems = isLoggedIn
+        ? [
+            { href: "/", label: "Home" },
+            { href: "/payment", label: "Payment" },
+            { href: "#", label: "Logout", onClick: onLogout },
+          ]
+        : [
+            { href: "/", label: "Home" },
+            { href: "/login", label: "Login" },
+            { href: "/registration", label: "Register" },
+          ];
+
     return (
         <div>
-            <button onClick={toggleVisibility} className='nav-button-mobile'>
+            <button onClick={toggleVisibility} className="nav-button-mobile">
                 Menu
                 {isVisible ? false : true}
             </button>
-            {isVisible && 
+            {isVisible && (
                 <div className={`overlay-${isAnimating ? 'shown' : 'hidden'}`}>
-                    <nav className='nav-container-mobile'>
+                    <nav className="nav-container-mobile">
                         <a href="/">
                             <img src="/home.png" alt="Home Logo" className="nav-logo" />
                         </a>
-                        <button onClick={toggleVisibility} className='nav-button-mobile'>
+                        <button onClick={toggleVisibility} className="nav-button-mobile">
                             Menu
                             {isVisible ? false : true}
                         </button>
                     </nav>
-                    <div className='mobile-menu'>
-                        <a className="nav-button" style={{animationDelay: '0.15s'}} href="/">Home</a>
-                        <a className="nav-button" style={{animationDelay: '0.2s'}} href="/login">Login</a>
-                        <a className="nav-button" style={{animationDelay: '0.25s'}} href="/registration">Register</a>
-                        <a className="nav-button" style={{animationDelay: '0.3s'}} href="/payment">Payment</a>
+                    <div className="mobile-menu">
+                        {menuItems.map((item, index) => (
+                            <a
+                                key={item.label}
+                                className="nav-button"
+                                style={{ animationDelay: `${0.15 + index * 0.05}s` }}
+                                href={item.href}
+                                onClick={(e) => {
+                                    if (item.onClick) {
+                                        e.preventDefault();
+                                        item.onClick();
+                                    }
+                                    toggleVisibility();
+                                }}
+                            >
+                                {item.label}
+                            </a>
+                        ))}
                     </div>
                 </div>
-            }
+            )}
         </div>
     );
 };
