+++ hotel/src/app/layout.js
@@ -3,98 +3,11 @@
 import { Inter } from "next/font/google";
 import { useState, useEffect } from 'react';
 import "./main.css";
+import ToggleComponent from "./components/mobileNav";
 
 const inter = Inter({ subsets: ["latin"] });
 
 // Mobile Navigation Toggle Component
-const ToggleComponent = ({ isLoggedIn, onLogout }) => {
-  // State for managing mobile menu open/close
-  const [isOpen, setIsOpen] = useState(false);
-
-  // Define all navigation items in an organized structure
-  const mobileMenuItems = {
-    alwaysShow: [
-      { href: "/", label: "Home" }
-    ],
-    loggedOut: [
-      { href: "/login", label: "Login" },
-      { href: "/registration", label: "Register" }
-    ],
-    loggedIn: [
-      { href: "/profile", label: "Profile" },
-      // Add logout as a menu item with special onClick handling
-      { href: "#", label: "Logout", onClick: onLogout }
-    ]
-  };
-
-  // Toggle mobile menu state
-  const toggleMenu = () => {
-    setIsOpen(!isOpen);
-  };
-
-  return (
-    <div className="container">
-      <div className="hamburger-menu">
-        <input
-          type="checkbox"
-          className="toggler"
-          checked={isOpen}
-          onChange={toggleMenu}
-        />
-        <div className="hamburger">
-          <div></div>
-        </div>
-        <div className={`menu ${isOpen ? 'open' : ''}`}>
-          <div>
-            <div>
-              {/* Always visible navigation items */}
-              {mobileMenuItems.alwaysShow.map((item) => (
-                <a
-                  key={item.href}
-                  href={item.href}
-                  onClick={() => setIsOpen(false)}
-                >
-                  {item.label}
-                </a>
-              ))}
-
-              {/* Conditional navigation based on authentication */}
-              {!isLoggedIn ? (
-                // Show login/register when logged out
-                mobileMenuItems.loggedOut.map((item) => (
-                  <a
-                    key={item.href}
-                    href={item.href}
-                    onClick={() => setIsOpen(false)}
-                  >
-                    {item.label}
-                  </a>
-                ))
-              ) : (
-                // Show authenticated menu items when logged in
-                mobileMenuItems.loggedIn.map((item) => (
-                  <a
-                    key={item.label}
-                    href={item.href}
-                    onClick={(e) => {
-                      if (item.onClick) {
-                        e.preventDefault();
-                        item.onClick();
-                      }
-                      setIsOpen(false);
-                    }}
-                  >
-                    {item.label}
-                  </a>
-                ))
-              )}
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
 
 // Main Layout Component
 export default function RootLayout({ children }) {
